requirements.txt
Flask==2.0.1
Pillow
opencv-python==4.5.3.56
numpy==1.21.2
gunicorn==20.1.0
requests==2.31.0
tqdm==4.66.1
